final-pjt-back/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
exceptiongroup==1.2.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
joblib==1.4.2
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyarrow==16.1.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.29.0
whitenoise==6.6.0
yarl==1.9.4
